[ISSUE-3] Refactor packages
